corrected comments and renamed reducer files
