+++ app/src/containers/Events/EventsList.js
@@ -3,7 +3,7 @@ import { Link } from 'react-router';
 import { ListGroup, ListGroupItem } from 'react-bootstrap';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import * as eventsActions from '../../redux/EventsReducers';
+import * as eventsActions from '../../redux/eventsReducers';
 import './eventsList.scss';
 
 class EventsList extends React.Component {
