+++ app/src/redux/eventsReducers.js
